Changed video generation to pipe Gource -> FFmpeg using FIFO
In an effort to speed up the video creation task and to reduce temporary disk usage,
this commit changes the `--output-ppm-stream` option to write to a FIFO (named pipe) file.
This file can be immediately streamed to the FFmpeg process to begin converting the video
without waiting for Gource to complete.

Also adjusted the FFmpeg process to send its stdout/stderr to files in the working directory.
Without flushing these file descriptors, the FFmpeg process could halt waiting for the buffer
to be cleared.  This has been observed on some projects with long video outputs (>10 minutes).
This could also be resolved by reducing the log verbosity, but there is plans to monitor
FFmpeg output to generate a progress bar.

- Added {% duration %} template tag to convert seconds to HH:mm:ss display
- Added helper method to get build duration (of background task); display on details page
- Consolidated 'project_details' and 'project_build_details' into one view
- Fixed older builds to display project options in readonly manner