Initial commit of user accounts and project roles; added private projects
This commit introduces using Django user accounts to authenticate access to Projects.
With this setup, only Project creators and assigned `ProjectMembers` can modify
project settings or request builds.  Currently these roles are limited to "developer"
and "maintainer"; either of which can edit a Project settings or request a new build.
You must also now have an account to create a new Project.

I also added an `is_public` flag to toggle hiding a Project from public listings.
In order to view the project, you must be assigned as a `ProjectMember`.  Currently
an assignment must be a "Developer" or "Maintainer", but I may add a new view-only role later.
To streamline this, I created a new Project `Manager` instance with a `filter_permissions()`
method that will filter the SQL scoped to the provided user (actor), as well as a custom
prefetch method to select the "latest" build.

To support user login, I added login/logout buttons to the navbar, and added a 'login.html'
template.  I would like for this login to be done using an inline popover from the navbar,
but having a dedicated login page is a good fallback.

Lastly, I converted the Django settings to use `django-split-settings`, which is a clean way
of separating out and optionally including settings files.