Added setting stubs for software path overrides; fixed Mercurial command name
Each software command used has an override setting to allow for a custom build
or install of the software to be used (vs. the OS default path).

    GIT_PATH            git
    MERCURIAL_PATH      hg
    GOURCE_PATH         gource
    FFMPEG_PATH         ffmpeg
    FFPROBE_PATH        ffprobe
    FFPLAY_PATH         ffplay

On some platforms this may be necessary as the provided package version may
be out of date.  In particular, FFmpeg has some bugs < 4.0 that affect
audio looping during mixing.

Fixed the `get_mercurial_path()` helper to look for `hg` command vs. non-existent `mercurial` command