Enabled experimental 'xvfb' support for headless Gource render
Gource requires a GUI in order to render its frames, but this may cause issues
running in some environments and uses up hardware resources.

This commit enables use of `Xvfb` (X virtual frame buffer) via a `xvfb-run`
command, which launches a virtual X11 buffer for Gource to output its frames
to.  When experimenting, this also lowers CPU usage during rendering and
speeds up the overall video task.  This can be enabled using new Django settings

    USE_XVFB = True
    XVFB_RUN_PATH = "/path/to/xvfb-run" # Optional

However, I have encountered a number of issues launching `xvfb-run` consistently
with Gource; about half the time it errors with a 'Segmentation fault' or

    gource: SDL initialization failed - No available video device

This is entirely sporadic, and waiting a new minutes to re-queue a new build
will often succeed.  It isn't clear if we need further teardown of the Xvfb
process upon success or not, but at the least we can usually detect when
the Gource process has aborted early so we don't launch an unnecessary `ffmpeg`
command.