Disable partial-clone/promisor config settings from remote downloads
When downloading remote Git repositories in order to generate a fresh Gource log,
we perform a "partial clone" using `--filter=blob:none` to avoid a large local
workspace when we only need the commit history/file changes.  In many versions
of Git this has not been an issue to perform a `git log` on these reduced checkouts.

However, on some newer Git versions (e.g. 2.25.1) when a partial clone is performed
and the `git log --raw` command is run, it will attempt to lazy fetch additional
artifacts from the origin.  This is potentially needed for the local repo to resolve
some `git diff` indexes when the `--raw` flag is used (without it, no network activity
is generated).  This lazy fetching is a problem for our application as a) we do not
want the addition file contents retrieved and b) the fetching is extremely slow as
it does not batch the same as a normal checkout.  This lazy request would cause
even relatively small repos to time out during the download process in our app.

This commit attempts to disable this partial clone fetching by modifying the
local checkout `.git/config` to remove the settings configuring this feature.

    [remote "origin"]
        promisor = true
        partialclonefilter = blob:none

With either of these configured, the partial repo attempts to fetch upstream
resources when any number of `git` command are run, particularly by Gource.
By removing them, it should disable the background fetching from occurring,
though it is possible that some diff history is not generated perfectly.
I have seen some stderr messages about not reading historical commits when
`git log --raw` is run, though I can't figure out which blobs cause it.
It may be that the options Gource uses prevents it from happening, IDK.

Not all versions of Git will store these settings under the `remote.origin`
section, but this is the location I observed on newer Git versions that caused
the background download issue.

Other changes:
- Added a `get_git_version()` helper to query current Git version in use
- Updated example `custom_settings.py`
- Updated README.md (dev setup/Xvfb usage)