Added 'ProjectMemberGroup' to manage project membership via Group
Similar to `ProjectMember`, a new `ProjectMemberGroup` uses the built-in
Group model to relate roles to a multiple users for a Project.
The main use case is for viewing private projects, which I also created
a new "viewer" role for members.

UI not fully ready to handle 'developer' vs 'maintainer' roles, but
a 'viewer' role shouldn't be capable of editing right now.