Moved default Gource video arguments to project option defaults
When running the Gource command, the shell arguments contained a list of
known "good" video display options for every video.  These options
were recently added to the whitelist of available options, so it makes
sense to make them mutable on each project for more customization.

This commit adds a data migration to populate all the current video
defaults on all Project's options, and adds a signal to automatically
create these when a new Project is instantiated.

- Fixed CSS on Project page to accommodate more label width
- Fixed Project options JS schema to accept Boolean values
- Fixed Project options JS to list "falsy" defaults (0, false) correctly
- Updated unit tests that were expecting empty `options` on creation